old/compute/requirements.txt
google-cloud-pubsub==2.28.0
google-cloud-compute==1.24.0
paramiko==3.5.0
google-cloud-storage==3.0.0
python-dotenv==1.0.1
google-api-python-client==2.160.0